requirements.txt
langchain
langchain-groq
pandas
streamlit
python-dotenv
langchain_core
langchain_openai